requirements.txt
Pillow
numpy
gpiozero
RPi.GPIO
evdev
termcolor
grove.i2c
redis
serial
psutil
inotify_simple
pyudev
sugarpi